+++ wp-content/plugins/appiappi-template-showcase/appiappi-template-showcase.php
@@ -21,6 +21,7 @@ require APPIAPPI_SHOWCASE_DIR . 'includes/cpt.php';
 require APPIAPPI_SHOWCASE_DIR . 'includes/taxonomy-meta.php';
 require APPIAPPI_SHOWCASE_DIR . 'includes/meta-boxes.php';
 require APPIAPPI_SHOWCASE_DIR . 'includes/shortcode.php';
+require APPIAPPI_SHOWCASE_DIR . 'includes/settings.php';
 
 /**
  * Featured images require post-thumbnails theme support, which the